requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.2.5
django-appconf==1.0.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2==2.9.8
psycopg2-binary==2.9.8
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
